@@ -6,32 +6,35 @@
 
     <properties>
         <project.main>de.funboyy.challenge.RandomDropsPlugin</project.main>
-        <maven.compiler.source>8</maven.compiler.source>
-        <maven.compiler.target>8</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
     </properties>
 
-    <groupId>de.funboyy.minecraft</groupId>
-    <artifactId>RandomDrops</artifactId>
-    <version>0.1.1-SNAPSHOT</version>
+    <groupId>de.funboyy</groupId>
+    <artifactId>random-drops</artifactId>
+    <version>0.2.0-SNAPSHOT</version>
 
     <name>RandomDrops</name>
     <inceptionYear>2022</inceptionYear>
 
     <organization>
         <name>Funboyy</name>
+        <url>https://github.com/Funboyy</url>
     </organization>
 
     <dependencies>
+        <!--suppress VulnerableLibrariesLocal -->
         <dependency>
             <groupId>org.spigotmc</groupId>
             <artifactId>spigot</artifactId>
-            <version>1.16.5-R0.1-SNAPSHOT</version>
+            <version>1.19.4-R0.1-SNAPSHOT</version>
             <scope>provided</scope>
         </dependency>
+
         <dependency>
             <groupId>org.projectlombok</groupId>
             <artifactId>lombok</artifactId>
-            <version>1.18.20</version>
+            <version>1.18.26</version>
             <scope>provided</scope>
         </dependency>
     </dependencies>
