@@ -4,11 +4,12 @@ import de.funboyy.challenge.command.ChallengeCommand;
 import de.funboyy.challenge.listener.ChallengeListener;
 import de.funboyy.challenge.listener.ProtectionListener;
 import de.funboyy.challenge.listener.RandomDropListener;
+import de.funboyy.challenge.tabcompleter.ChallengeTabCompleter;
 import de.funboyy.challenge.utils.Timer;
 import lombok.Getter;
 import net.md_5.bungee.api.ChatColor;
 import net.md_5.bungee.api.ChatMessageType;
-import net.md_5.bungee.api.chat.TextComponent;
+import net.md_5.bungee.api.chat.ComponentBuilder;
 import org.bukkit.Bukkit;
 import org.bukkit.command.PluginCommand;
 import org.bukkit.plugin.java.JavaPlugin;
@@ -33,8 +34,10 @@ public class RandomDropsPlugin extends JavaPlugin {
         Bukkit.getPluginManager().registerEvents(new RandomDropListener(), this);
 
         final PluginCommand command = getCommand("challenge");
+
         if (command != null) {
             command.setExecutor(new ChallengeCommand());
+            command.setTabCompleter(new ChallengeTabCompleter());
         }
     }
 
@@ -47,11 +50,11 @@ public class RandomDropsPlugin extends JavaPlugin {
 
     private void timer() {
         Bukkit.getScheduler().runTaskTimerAsynchronously(this, () -> Bukkit.getOnlinePlayers().forEach(player -> {
-            final TextComponent text = new TextComponent(this.timer.toString());
-            text.setColor(ChatColor.GOLD);
-            text.setBold(true);
+            final ComponentBuilder builder = new ComponentBuilder("")
+                    .append(this.timer.toString()).color(ChatColor.GOLD).bold(true);
 
-            player.spigot().sendMessage(ChatMessageType.ACTION_BAR, text);
+            player.spigot().sendMessage(ChatMessageType.ACTION_BAR, builder.create());
         }), 0, 10);
     }
+
 }
